ingest/requirements.txt
fastapi==0.115.8
jsonschema==4.23.0
opentelemetry-api==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-instrumentation-fastapi==0.51b0
paho-mqtt==2.1.0
prometheus-client==0.21.1
psycopg[binary]==3.2.6
python-dotenv==1.0.1
uvicorn==0.34.0
